requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pytest
mecab-python3
pgmpy
pandas
numpy
unidic-lite